* device/lib/_gptrget.c, * device/lib/_gptrput.c: updated comment * device/lib/calloc.c, * device/lib/free.c, * device/lib/malloc.c, * device/lib/realloc.c: added LGPL, made them reentrant-safe * src/SDCCcse.c (cseBBlock), * src/SDCCicode.c (printOperand, geniCodeArray), * src/SDCCicode.h (struct operand): fixed bug 868103 * support/regression/tests/bug-868103.c: added * src/SDCCast.c (searchLitOp), * src/SDCCcse.h (struct cseDef), * src/SDCCglue.c (printIvalArray, spacesToUnderscores), * src/SDCCicode.h (struct operand), * src/SDCCsymt.h (struct sym_link), * src/avr/gen.c (hasInc), * src/ds390/gen.c (hasInc), * src/hc08/gen.c (genPlusIncr, hasInc), * src/mcs51/gen.c (hasInc), * src/pic16/glue.c (pic16_printIvalChar), * src/pic16/ralloc.c (regWithIdx), * src/xa51/gen.c (hasInc) : removed warnings * src/SDCCast.c (createBlock): added comment ??? * src/hc08/ralloc.c: updated comments
git-svn-id: https://sdcc.svn.sourceforge.net/svnroot/sdcc/trunk/sdcc@3482 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ src/SDCCsymt.h
@@ -224,7 +224,7 @@ typedef struct sym_link
       unsigned intrtn:1;	/* this is an interrupt routin */
       unsigned rbank:1;		/* seperate register bank     */
       unsigned intno;		/* 1=Interrupt svc routine    */
-      unsigned regbank;		/* register bank 2b used      */
+      short    regbank;		/* register bank 2b used      */
       unsigned builtin;		/* is a builtin function      */
       unsigned javaNative;    	/* is a JavaNative Function (TININative ONLY) */
       unsigned overlay;    	/* force parameters & locals into overlay segment */
